+++ rust/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### NEXT
 
+- `DirectDataProducer.send()`: Add `ignored_subchannel` optional argument ([PR #XXXX](https://github.com/versatica/mediasoup/pull/XXXX)).
+
 ### 0.24.3
 
 - `DataProducerOptions`: make `new_pipe_transport()` public ([PR #1876](https://github.com/versatica/mediasoup/pull/1876)).